[refactor] s3 관련 서비스 디렉토리 변경
