+++ src/main/java/ctrlS/totori/book/dto/response/BookPageResponse.java
@@ -1,7 +1,6 @@
 package ctrlS.totori.book.dto.response;
 
 import ctrlS.totori.book.entity.BookPage;
-import ctrlS.totori.book.service.S3ImageStorageService;
 
 import java.util.List;
