+++ src/main/java/ctrlS/totori/book/service/image/ImageStorageService.java
@@ -1,4 +1,4 @@
-package ctrlS.totori.book.service;
+package ctrlS.totori.book.service.image;
 
 public interface ImageStorageService {
     String uploadImage(byte[] imageBytes, String fileName);
