+++ src/main/java/ctrlS/totori/book/service/image/PageImageAsyncService.java
@@ -1,5 +1,6 @@
-package ctrlS.totori.book.service;
+package ctrlS.totori.book.service.image;
 
+import ctrlS.totori.book.service.StableDiffusionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
